refactor: make packages to organize structure
