+++ wallet/src/main/kotlin/net/thechance/wallet/api/dto/balance/BalanceResponse.kt
@@ -1,4 +1,4 @@
-package net.thechance.wallet.api.dto
+package net.thechance.wallet.api.dto.balance
 
 data class BalanceResponse(
     val balance: Double,
